parse light done
