+++ Makefile
@@ -28,8 +28,8 @@ SRC = minirt.c \
 	  create_world.c set_material.c intersect_world_with_ray.c prepare_computations.c shade_hit.c \
 	  color_at.c view_transformation.c create_camera.c ray_for_pixel.c render.c intersect_sphere.c \
 	  intersect_plane.c intersect_cylinder.c create_plane.c create_cylinder.c cylinder_setters.c \
-	  shadows.c render_image.c endwith.c parse_ambient.c parse_camera.c parse_plane.c parse_utils.c \
-	  parse_cylinder.c parse_light.c parse_sphere.c parse_functions_table.c parse_tokens_into_world.c caps.c \
+	  shadows.c render_image.c endwith.c parse_ambient_and_light.c parse_camera.c parse_plane.c parse_utils.c \
+	  parse_cylinder.c parse_sphere.c parse_functions_table.c parse_tokens_into_world.c caps.c \
 	  check_type_identifiers.c file_validation.c open_file.c validate_line.c tokenizer.c 
 
 VPATH = ./src ./src/math/canvas/ ./src/math/tuples/ ./src/math/ppm/ ./src/math/matrices \
