+++ test.rt
@@ -21,7 +21,7 @@ cy 50.0,0.0,20.6 0,0,1.0 14.2 21.42 10,0,255
 cy 50.0,0.0,20.6 0,0,1.0 14.2 21.42 10,0,255
 cy 50.0,0.0,20.6 0,0,1.0 14.2 21.42 10,0,255
 cy 50.0,0.0,20.6 0,0,1.0 14.2 21.42 10,0,255
-L -41,0,30 0.7
+L -40.0,50.0,0.0 2.6
 cy 50.0,0.0,20.6 0,0,1.0 14.2 21.42 10,0,255
 cy 50.0,0.0,20.6 0,0,1.0 14.2 21.42 10,0,255
 cy 50.0,0.0,20.6 0,0,1.0 14.2 21.42 10,0,255
